pyproject.toml
[project]
name = "frbsf-chart-builder"
version = "0.1.0"
description = "FRBSF Chart Builder - Economic chart creation and customization tool"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "opencv-python-headless>=4.9.0",
    "boto3>=1.34.0",
    "reportlab>=4.1.0",
    "pyyaml>=6.0",
    "python-multipart>=0.0.9",
    "aiosqlite>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "hypothesis>=6.98.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
